server/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
dm-tree==0.1.8
flake8==7.0.0
flake8-bugbear==24.2.6
flatbuffers==24.3.7
fonttools==4.50.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mypy-extensions==1.0.0
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pandas-stubs==2.2.1.240316
pillow==10.2.0
protobuf==4.25.3
psycopg==3.1.18
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-surprise==1.1.3
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
types-cryptography==3.3.23.2
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
